시설 안내 관련 타입 정리
