+++ components/facilities/FacilitiesRow.tsx
@@ -19,10 +19,8 @@ export default function FacilitiesRow({
   return (
     <article className="text-neutral-700 font-noto flex my-[1.2rem] flex-row items-start justify-between break-all h-40">
       <div className="flex flex-col w-[35.5rem] h-auto">
-        <Link href="" className="hover:underline">
-          <h3 className="text-neutral-800 text-md font-bold mb-[.69rem] leading-5">{name}</h3>
-        </Link>
-        <Link href="" className="hover:cursor-pointer flex flex-col gap-1">
+        <h3 className="text-neutral-800 text-md font-bold mb-[.69rem] leading-5">{name}</h3>
+        <div className="flex flex-col gap-1">
           <div className="text-sm font-normal leading-[1.63rem]">
             <HTMLViewer htmlContent={description} />
           </div>
@@ -32,11 +30,11 @@ export default function FacilitiesRow({
             </span>
             <p className="text-sm font-normal leading-[1.63rem]">{location}</p>
           </div>
-        </Link>
+        </div>
       </div>
-      <Link href="" className="w-60 h-40 relative">
+      <div className="w-60 h-40 relative">
         <Image alt="대표 이미지" src={imageURL} fill sizes="10rem" priority />
-      </Link>
+      </div>
     </article>
   );
 }
