--- types/facilities.ts
@@ -1,8 +1,0 @@
-export interface FacilitiesResponse {
-  facilitiesList: {
-    name: string;
-    description: string;
-    location: string;
-    imageURL: string;
-  }[];
-}
